Add and use the FixedMultiColumn layout
This has been sitting around for quite a while, because I feel like it
should be combined with FixedColumn, but I'm not sure how best to do
that without breaking compatibility for people. I'm just checking it
in to my configs for now, and I'll decide how to get it upstream
later.